This commit was generated by cvs2svn to compensate for changes in r13031, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@13032 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/tape/t64.c
@@ -297,7 +297,7 @@ int t64_read(t64_t *t64, BYTE *buf, size_t size)
     int amount;
 
     if (t64 == NULL || t64->fd == NULL || t64->current_file_number < 0
-        || size < 0)
+        || (int)size < 0)
         return -1;
 
     if (size == 0)
